fead (ElectronicTable)

+++ fabi/src/main/java/com/ninotech/fabi/controleur/adapter/PhysicalAdapter.java
@@ -105,7 +105,7 @@ public class PhysicalAdapter extends RecyclerView.Adapter<PhysicalAdapter.MyView
             mTitleTextView.setText(physical.getTitle());
             mDateInitTextView.setText(physical.getDateStart());
             mDateFinaleTextView.setText(physical.getDateEnd());
-            mProgressBar.setProgress(physical.getPercentage());
+            mProgressBar.setProgress((int) physical.getPercentage());
             mPercentageTextView.setText(String.valueOf(physical.getPercentage()) + "%");
         }
     }
